history: Move history methods from BaseCli to HistoryCommand
